chore: move ssmr correctly to first party where it belongs
