+++ src/data/lancer/lcp/ssmr part 1/lcp_manifest.json
