+++ src/data/lancer/lcp/ssmr part 1/weapons.json
